+++ src/Clone.js
@@ -233,7 +233,10 @@ export class DocumentCloner {
         }
 
         for (let child = node.firstChild; child; child = child.nextSibling) {
-            if (child.nodeType !== Node.ELEMENT_NODE || child.nodeName !== 'SCRIPT') {
+            if (
+                child.nodeType !== Node.ELEMENT_NODE ||
+                (child.nodeName !== 'SCRIPT' && !child.hasAttribute('data-html2canvas-ignore'))
+            ) {
                 if (!this.copyStyles || child.nodeName !== 'STYLE') {
                     clone.appendChild(this.cloneNode(child));
                 }