+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### Changelog ###
 
 #### v1.0.0-alpha2 - TBD ####
+ * Fix `data-html2canvas-ignore` attribute
  * Fix decimal `letter-spacing` values
 
 #### v1.0.0-alpha1 - 5.12.2017 ####